refactor(core): extract password generator into services/
The 694-line commands/generator.rs was a service file pretending to be
a command file. The Tauri command handlers were 5-line passthroughs to
private `_value` helpers in the same module, and the helpers
(validate_min_requirements, enforce_minimums, filter_charset, the EFF
diceware wordlist, charset constants) were the actual implementation.

Moves all production code (~285 lines) to services/generator.rs:
- PasswordGeneratorOptions / PassphraseGeneratorOptions
- GeneratedPassword / GeneratedPassphrase
- generate_password / generate_passphrase (renamed from *_value)
- validate_min_requirements, enforce_minimums, filter_charset
- charset constants, AMBIGUOUS_CHARS, WORDLIST

Tests (~340 lines) move alongside the code they exercise.

commands/generator.rs is now 23 lines: re-exports the DTOs from the
service and declares two thin `#[tauri::command]` handlers that
delegate. The IPC surface is unchanged — `generate_password` and
`generate_passphrase` keep their names, parameters, and return shapes.
CLAUDE.md's "commands are thin handlers that delegate to services" is
now true for this module.

No behavioral changes. Frontend (PasswordGenerator.tsx, etc.) is
unaffected because the camelCase-serialized JSON contract is identical.
