+++ src-tauri/src/services/mod.rs
@@ -3,6 +3,7 @@
 pub mod auto_lock;
 pub mod clipboard;
 pub mod crypto;
+pub mod generator;
 pub mod kdbx;
 pub mod secure_storage;
 pub mod settings;
